chore: restructure eventrecorder package
